python-servers/requirements.txt
flask
flask-cors
transformers
torch
numpy
pandas
scipy
matplotlib
pinecone-client
openai-whisper
youtube-transcript-api
pypdf2
pytesseract
opencv-python
Pillow
python-dotenv
sentencepiece
sentence-transformers
